requirements.txt
openai
dbus-python
PyGObject
regex